+++ src/main/java/com/planbase/pdf/layoutmanager/lineWrapping/LineWrapped.kt
@@ -20,6 +20,7 @@
 
 package com.planbase.pdf.layoutmanager.lineWrapping
 
+import com.planbase.pdf.layoutmanager.attributes.DimAndPages
 import com.planbase.pdf.layoutmanager.pages.RenderTarget
 import com.planbase.pdf.layoutmanager.utils.Dim
 import com.planbase.pdf.layoutmanager.utils.Coord
@@ -54,7 +55,7 @@ interface LineWrapped {
      * @return the adjusted Dim which may include extra (vertical) spacing required to nudge some items onto the next
      * page so they don't end up in the margin or off the page.
      */
-    fun render(lp: RenderTarget, topLeft: Coord, reallyRender:Boolean): Dim
+    fun render(lp: RenderTarget, topLeft: Coord, reallyRender:Boolean): DimAndPages
 
     /**
      * Sends the underlying object to PDFBox to be drawn. Use the other render() method with reallyRender=false
@@ -66,19 +67,20 @@ interface LineWrapped {
      * @return the adjusted Dim which may include extra (vertical) spacing required to nudge some items onto the next
      * page so they don't end up in the margin or off the page.
      */
-    fun render(lp: RenderTarget, topLeft: Coord) = render(lp, topLeft, true)
+    fun render(lp: RenderTarget, topLeft: Coord): DimAndPages = render(lp, topLeft, true)
 
-    object ZeroLineWrapped: LineWrapped {
-        override val dim: Dim = Dim.ZERO
-
-        override val ascent: Float = 0f
-
-//        override val descentAndLeading: Float = 0f
-
-        override val lineHeight: Float = 0f
-
-        override fun render(lp: RenderTarget, topLeft: Coord, reallyRender: Boolean): Dim = dim
-    }
+//    object ZeroLineWrapped: LineWrapped {
+//        override val dim: Dim = Dim.ZERO
+//
+//        override val ascent: Float = 0f
+//
+////        override val descentAndLeading: Float = 0f
+//
+//        override val lineHeight: Float = 0f
+//
+//        override fun render(lp: RenderTarget, topLeft: Coord, reallyRender: Boolean) = DimAndPages(dim, lp.)
+//
+//    }
 //    companion object {
 //
 //        fun preWrappedLineWrapper(item: LineWrapped) = object : LineWrapper {